refactor: Lookup Orders components
